requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord-webhook==1.0.0
Flask==2.2.2
idna==3.4
importlib_metadata==8.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.1.1
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.23.0
